deleted files scripts
