+++ prisma/schema.prisma
@@ -24,7 +24,7 @@ model UserProfile {
   myaddress                String
   secondAddress            String?
   createdAt                DateTime             @default(dbgenerated("now()"))
-  isActive                 Boolean
+  isActive                 Boolean              @default(true)
   createdGroups            Group[]              @relation("CreatedBy") // list of created groups
   groupMember              GroupMember[]        @relation("groupMember") // list of groups that user belongs to
   wishlists                Wishlist[] // wishlists created by this user
