--- src/backend/prisma.js
@@ -1,5 +1,0 @@
-import { PrismaClient } from '@prisma/client'
-
-export const prisma = global.prisma || new PrismaClient()
-
-if (process.env.NODE_ENV !== 'production') global.prisma = prisma
