+++ src/backend/supabase/.gitignore
